requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--no-binary :all:

certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
django-anymail==6.0
django-cors-headers==2.4.0
django-enumfields==1.0.0
django-environ==0.4.5
django-filter==2.1.0
django-helusers==0.4.2
django-js-asset==1.1.0    # via django-mptt
django-mailer==1.2.5
django-mptt==0.9.1        # via django-munigeo
django-munigeo==0.3.2
django-parler-rest==2.0
django-parler==1.9.2
django==2.1.3
djangorestframework-gis==0.14
djangorestframework==3.9.0
drf-oidc-auth==0.9        # via django-helusers
future==0.17.1            # via pyjwkest
holidays==0.9.9
idna==2.8                 # via requests
jinja2==2.10
lockfile==0.12.2          # via django-mailer
markupsafe==1.1.0         # via jinja2
psycopg2-binary==2.7.6.1
pycryptodomex==3.7.2      # via pyjwkest
pyjwkest==1.4.0           # via drf-oidc-auth
python-dateutil==2.7.5    # via holidays
pytz==2018.7              # via django
pyyaml==3.13              # via django-munigeo
raven==6.9.0
requests-cache==0.4.13    # via django-munigeo
requests==2.21.0          # via django-anymail, django-helusers, django-munigeo, pyjwkest, requests-cache
six==1.12.0               # via django-anymail, django-munigeo, holidays, pyjwkest, python-dateutil
urllib3==1.24.1           # via requests
